requirements.txt
asgiref==3.8.1
begins==0.9
Django==5.1.6
django-browser-reload==1.18.0
django-debug-toolbar==5.0.1
django-erd==1.0
django-extensions==3.2.3
django-resized==1.0.3
djangorestframework==3.15.2
pillow==11.1.0
pydot==3.0.4
pyparsing==3.2.1
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.3
djangorestframework-simplejwt==5.5.0
pydantic==2.11.3
ollama==0.4.8
requests==2.32.3
